### 2019-08-09 #1633 @whedon
[PRE REVIEW]: PressureDrop.jl: Pressure traverses for oil & gas wells
**Submitting author:** @jnoynaert (<a href="http://orcid.org/0000-0002-2986-0376">Jared Noynaert</a>)
**Repository:** <a href="https://github.com/jnoynaert/PressureDrop.jl" target ="_blank">https://github.com/jnoynaert/PressureDrop.jl</a>
**Version:** v1.0.1
