### 2019-08-09 #1633 @arfon
Thanks for your submission to JOSS @jnoynaert. Please could you remove the backticks around the citations you have in your `paper.md` so Pandoc can pick them up?
