+++ dev/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"#PressureDrop.jl-Documentation-1","page":"Overview","title":"PressureDrop.jl Documentation","text":"","category":"section"},{"location":"#","page":"Overview","title":"Overview","text":"CurrentModule = PressureDrop","category":"page"},{"location":"#","page":"Overview","title":"Overview","text":"PressureDrop.jl is a Julia package for computing multiphase pressure profiles for gas lift optimization of oil & gas wells.","category":"page"},{"location":"#","page":"Overview","title":"Overview","text":"Outlet-referenced models for producing wells using non-coupled temperature gradients are currently supported.","category":"page"},{"location":"#","page":"Overview","title":"Overview","text":"note: Note\nNote that all inputs and calculations are in U.S. field units.","category":"page"},{"location":"#Contents-1","page":"Overview","title":"Contents","text":"","category":"section"},{"location":"#","page":"Overview","title":"Overview","text":"Pages = [\r\n  \"core.md\",\r\n  \"utilities.md\",\r\n  \"plotting.md\",\r\n  \"correlations.md\",\r\n  \"pvt.md\",\r\n  \"valves.md\"\r\n]\r\nDepth = 2","category":"page"},{"location":"core/#Core-functionality-1","page":"Core functions","title":"Core functionality","text":"","category":"section"},{"location":"core/#","page":"Core functions","title":"Core functions","text":"Pages = [\"core.md\"]\r\nDepth = 3","category":"page"},{"location":"core/#","page":"Core functions","title":"Core functions","text":"using PressureDrop\r\n\r\nsurveyfilepath = joinpath(dirname(dirname(pathof(PressureDrop))), \"test/testdata/Sawgrass_9_32/Test_survey_Sawgrass_9.csv\")\r\n\r\nvalvefilepath = joinpath(dirname(dirname(pathof(PressureDrop))), \"test/testdata/valvedata_wrappers_1.csv\")","category":"page"},{"location":"core/#Creating-and-updating-models-1","page":"Core functions","title":"Creating and updating models","text":"","category":"section"},{"location":"core/#","page":"Core functions","title":"Core functions","text":"Model definitions are created and stored as WellModel objects. Although all of the functionality of this package is exposed as pure functions, mutating and copying WellModels is a much easier way to track and iterate on parameter sets.","category":"page"},{"location":"core/#Wellbores-1","page":"Core functions","title":"Wellbores","text":"","category":"section"},{"location":"core/#","page":"Core functions","title":"Core functions","text":"The key component required for the pressure drop calculations is a Wellbore object that defines the flow path in terms of directional survey points (measured depth, inclination, and true vertical depth) and tubular inner diameter.","category":"page"},{"location":"core/#","page":"Core functions","title":"Core functions","text":"Wellbore objects can be constructed from arrays, or from CSV files with read_survey, which includes some optional convenience arguments to change delimiters, skip lines, or truncate the survey. Tubing IDs do not have to be uniform and can be specified segment to segment.","category":"page"},{"location":"core/#","page":"Core functions","title":"Core functions","text":"examplewell = read_survey(path = surveyfilepath, id = 2.441, maxdepth = 6500) #an outlet point at 0 MD is added if not present","category":"page"},{"location":"core/#Valve-designs-1","page":"Core functions","title":"Valve designs","text":"","category":"section"},{"location":"core/#","page":"Core functions","title":"Core functions","text":"GasliftValves objects define the valve strings in terms of measured run depth, test rack opening p
[…]
