### 2019-08-11 #1633 @oliviaguest
@mancellin are you interested and able to review *[PressureDrop.jl: Pressure traverses for oil & gas wells](https://github.com/jnoynaert/PressureDrop.jl)* for [JOSS](https://joss.theoj.org)?
