### 2019-08-12 #1633 @labarba
And once you start the review, close this issue, and go to the Review issue to post a welcoming message to get a ping out to the reviewers. I use something like:

> 👋 <rev1> <rev2> — This is where the action happens. You both have a review checklist to work your way through. Feel free to post questions here and open review issues in the submission repository (with a link to this issue).
