### 2019-08-12 #1633 @danielskatz
You may want to browse https://joss.readthedocs.io/en/latest/editing.html as well.

And if you want to ask something of the associate editor in chief on duty, please ping ```@openjournals/joss-eics``` rather than a specific person
