### 2019-08-12 #1642 @labarba
Hi everybody! If you open a new issue in the [submission repository](https://github.com/jnoynaert/PressureDrop.jl), it's nice if you link to this very issue thread here, so we get a cross-link from GitHub.
