### 2019-08-23 jnoynaert/PressureDrop.jl#3 @mancellin
Questions / Suggestions (JOSS review)
Hi @jnoynaert
Here are a few friendly suggestions that came to my mind while reviewing your package for the JOSS.

- The documentation would be more friendly to beginners if it included an example of `surveyfilepath` and `valvefilepath`.
- I could not easily find information about the theory behind your code: which equations are being solved and the method used. Adding a short summary of the theory in the documentation could be nice. For instance, from reading the code, it seems that the core functions of the code are using some Picard fixed point iterative method, am I right ?
- About this iterative method, a possible improvement would be to call a dedicated Julia library. It would make the code more powerful for advanced users and would fit within your goals of modularity and composability.
- The JOSS paper might be a good place to cite the other packages your code is relying on. [Gadfly's documentation](http://gadflyjl.org/stable/) suggests a citation for their package.
