### 2019-08-23 jnoynaert/PressureDrop.jl#3 @jnoynaert
Thanks! File paths are exactly the sort of thing that would trip up a lot of the target users.

I'll add some reference to the basic theory; you're correct that the solution is just fixed-point iteration. The level of effort to utilize NLPsolve.jl for the solutions should be pretty reasonable, so I'll try that out in a test branch.
