### 2019-08-27 d20b461
merge benchmarks from NLsolve test
