+++ Project.toml
@@ -13,7 +13,8 @@ Requires = "ae029012-a4dd-5104-9daa-d747884805df"
 julia = "1.0"
 
 [extras]
+BenchmarkTools = "6e4b80f9-dd63-53aa-95a3-0cdb28fa8baf"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["Test"]
+test = ["Test", "BenchmarkTools"]
