### 2019-08-27 bb5a516
changed dp_est to update with each segment (improved total time ~10% across integration benchmarks)
