+++ test/test_integration_scenario.jl
@@ -117,7 +117,7 @@ compare_tolerance = 40 #every scenario except the high-rate scenario can take a 
 for index in 2:length(scenarios)+1
     println("Comparing scenario ", index-1, " on index ", index)
     @test all( abs.(test_results[1:hz_index,index] .- matched_example[1:hz_index,index])  .<= compare_tolerance )
-    println("Max difference :", maximum(abs.(test_results[1:hz_index,index] .- matched_example[1:hz_index,index])))
+    println("Max difference : ", maximum(abs.(test_results[1:hz_index,index] .- matched_example[1:hz_index,index])))
 end
 
 end #testset for B&B scenarios
@@ -155,7 +155,7 @@ compare_tolerance = 85 #lloser tolerance to H&B due to wide range of methods to 
 for index in 2:length(scenarios)+1
     println("Comparing scenario ", index-1, " on index ", index)
     @test all( abs.(test_results[1:hz_index,index] .- matched_example[1:hz_index,index])  .<= compare_tolerance )
-    println("Max difference :", maximum(abs.(test_results[1:hz_index,index] .- matched_example[1:hz_index,index])))
+    println("Max difference : ", maximum(abs.(test_results[1:hz_index,index] .- matched_example[1:hz_index,index])))
 end
 
 end #H&B testset
