### 2019-08-27 jnoynaert/PressureDrop.jl#3 @mancellin
> File paths are exactly the sort of thing that would trip up a lot of the target users.

To clarify, I was thinking of providing examples of the content of the files. Although, you could also add examples of the paths themselves.
