### 2019-08-28 e021fca
cleaned up segment calculation; improved total time addtl ~5-10%
