+++ src/PressureDrop.jl
@@ -65,56 +65,28 @@ end
 
 #%% core functions
 """
-`calculate_pressuresegment_topdown(<arguments>)`
+`calculate_pressuresegment(<arguments>)`
 
 Pressure inputs are in **psia**.
 
 Helper function to calculate the pressure drop for a single pipe segment, using an outlet-referenced approach.
 
-Method:
+Method (fixed-point iteration to account for pressure dependence of fluid PVT properties) :
 1. Generates PVT properties for the average conditions in the segment for a given estimate for the pressure drop in the second.
 2. Calculates a new estimated pressure drop using the PVT properties.
 3. Compares the original and new pressure drop estimates to validate the stability of the estimate.
 4. Iterate through steps 1-3 until a stable estimate is found (indicated by obtaining a difference between pre- and post-PVT that is within the given error tolerance).
 
 """
-function calculate_pressuresegment_topdown(pressurecorrelation::Function, p_initial, dp_est, t_avg,
-                                            md_initial, md_end, tvd_initial, tvd_end, inclination, id, roughness,
+function calculate_pressuresegment(pressurecorrelation::Function, p_initial, dp_est, t_avg,
+                                            dh_md, dh_tvd, inclination, uphill_flow, id, roughness,
                                             q_o, q_w, GLR, R_b, APIoil, sg_water, sg_gas, molFracCO2, molFracH2S,
-                                            pseudocrit_pressure_correlation::Function, pseudocrit_temp_correlation::Function, Z_correlation::Function,
+                                            Z_correlation::Function, P_pc, T_pc,
                                             gas_viscosity_correlation::Function, solutionGORcorrelation::Function, bubblepoint, oilVolumeFactor_correlation::Function, waterVolumeFactor_correlation::Function,
                                             dead_oil_viscosity_correlation::Function, live_oil_viscosity_correlation::Function, frictionfactor::Function, error_tolerance = 0.1)
 
-    dh_md = md_end - md_initial
-    dh_tvd = tvd_end - tvd_initial
-    p_avg = p_initial + dp_est/2
-    uphill_flow = inclination <= 90.0
-
-
-    P_pc = pseudocrit_pressure_correlation(sg_gas, molFracCO2, molFracH2S)
-    _, T_pc, _ = pseudocrit_temp_correlation(sg_gas, molFracCO2, molFracH2S)
-    Z = Z_correlation(P_pc, T_pc, p_avg, t_avg)
-    ρ_g = gasDensity_insitu(sg_gas, Z, p_avg, t_avg)
-    B_g = gasVolumeFactor(p_avg, Z, t_avg)
-    μ_g = gas_viscosity_correlation(sg_gas, p_avg, t_avg, Z)
-    R_s = solutionGORcorrelation(APIoil, sg_gas, p_avg, t_avg, R_b, bubblepoint)
-    v_sg = gasvelocity_superficial(q_o, q_w, GLR, R_s, id, B_g)
-    B_o = oilVolumeFactor_correlation(APIoil, sg_gas, R_s, p_avg, t_avg)
-    B_w = waterVolumeFactor_correlation(p_avg, t_avg)
-    v_sl = liquidvelocity_superficial(q_o, q_w, id, B_o, B_w)
-    ρ_l = mixture_properties_simple(q_o, q_w, oilDensity_insitu(APIoil,  sg_gas,  R_s,  B_o), waterDensity_insitu(sg_water, B_w))
-    σ_l = mixture_properties_simple(q_o, q_w, gas_oil_interfacialtension(APIoil, p_avg, t_avg), gas_water_interfacialtension(p_avg, t_avg))
-    μ_oD = dead_oil_viscosity_correlation(APIoil, t_avg)
[…]
