### 2019-08-28 b01a516
restructured casing segment calc & matched all internal fn tests to new signatures
