### 2019-08-28 jnoynaert/PressureDrop.jl#3 @jnoynaert
That would, indeed, be much more useful than a file path example...!

I really liked your idea of of using a dedicated solver library, so I tested using NLsolve.jl to calculate the solution of each pressure segment:
- This seemed like the most reasonable dedicated package for this purpose, other than treating it as optimization problem via JuMP.jl or similar.
- I ended up dropping the change, as using Anderson acceleration with a variety of tweaks actually made things both slower and less stable, while just using the NLsolve equivalent of naive fixed-point (Anderson with 0 memory) just made things slower.
    - After investigating this, I think the issue is that across all test cases, the maximum number of iterations to converge for a segment turned out to be 5 (more typically 2-3). Thus the overhead of more advanced methods doesn't pay out, and wrappers over the naive method don't offer any improvement either.
- The very positive outcome of this line of investigation was that in the process of refactoring functions to match NLsolve's expected input, a few performance easy performance optimizations became obvious, so now most cases will run 15-20% faster :)

When I have time I'll implement the changes to the paper & documentation that you suggested, and tag you once that's complete.
