### 2019-08-29 3d77dcc
updated paper & documentation
