+++ paper/paper.bib
@@ -47,3 +47,41 @@
   year={1980},
   publisher={American Society of Mechanical Engineers}
 }
+
+@misc{Jones:2019,
+  author       = {Daniel C. Jones and
+                  Ben Arthur and
+                  Tamas Nagy and
+                  Mattriks and
+                  Shashi Gowda and
+                  Godisemo and
+                  Tim Holy and
+                  Andreas Noack and
+                  Avik Sengupta and
+                  Darwin Darakananda and
+                  Adam B and
+                  Simon Leblanc and
+                  Iain Dunning and
+                  Keno Fischer and
+                  David Chudzicki and
+                  Morten Piibeleht and
+                  Yichao Yu and
+                  Tom Breloff and
+                  Dave Kleinschmidt and
+                  Alex Mellnik and
+                  john verzani and
+                  inkyu and
+                  Mike J Innes and
+                  Joey Huchette and
+                  Katharine Hyatt and
+                  Jared Forsyth and
+                  Gio Borje and
+                  Elliot Saba and
+                  Calder Coalson and
+                  Artem Pelenitsyn},
+  title        = {GiovineItalia/Gadfly.jl: v1.1.0},
+  month        = jul,
+  year         = 2019,
+  doi          = {10.5281/zenodo.3337394},
+  url          = {https://doi.org/10.5281/zenodo.3337394}
+}
