+++ test/test_casingcalcs.jl
@@ -4,7 +4,7 @@ include("../src/casingcalculations.jl")
     sg_gas = 0.7
     P_pc = HankinsonWithWichertPseudoCriticalPressure(sg_gas, 0, 0)
     _, T_pc, _ = HankinsonWithWichertPseudoCriticalTemp(sg_gas, 0, 0)
-    ΔP = casing_pressuresegment(300, 10, (100+180)/2, 4500, sg_gas, KareemEtAlZFactor, P_pc, T_pc)
+    ΔP = calculate_casing_pressuresegment(300, 10, (100+180)/2, 4500, sg_gas, KareemEtAlZFactor, P_pc, T_pc)
     @test ΔP ≈ 332.7-300 atol = 1
 end
 
