+++ docs/Project.toml
@@ -2,6 +2,7 @@
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
 Compose = "a81c6b42-2e10-5240-aca2-a61377ecd94b"
 Gadfly = "c91e804a-d5a3-530f-b6f0-dfbca275c004"
+PrettyTables = "08abe8d2-0d0c-5749-adfa-8a2ac140af0d"
 
 [compat]
 Documenter = "~0.22.4"

[…]
