+++ v1.0.2/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"#PressureDrop.jl-Documentation-1","page":"Overview","title":"PressureDrop.jl Documentation","text":"","category":"section"},{"location":"#","page":"Overview","title":"Overview","text":"CurrentModule = PressureDrop","category":"page"},{"location":"#","page":"Overview","title":"Overview","text":"PressureDrop.jl is a Julia package for computing multiphase pressure profiles for gas lift optimization of oil & gas wells.","category":"page"},{"location":"#","page":"Overview","title":"Overview","text":"Outlet-referenced models for producing wells using non-coupled temperature gradients are currently supported. ","category":"page"},{"location":"#","page":"Overview","title":"Overview","text":"note: Note\nNote that all inputs and calculations are in U.S. field units.","category":"page"},{"location":"#Overview-1","page":"Overview","title":"Overview","text":"","category":"section"},{"location":"#","page":"Overview","title":"Overview","text":"The pressure traverse along the producing flow path of an oil well (or its bottomhole pressure at the interface between the well and the reservoir) is critical information for many production and reservoir engineering workflows. This can be expensive to measure directly in many conditions, so 1D pressure and temperature models are frequently used to infer producing conditions.","category":"page"},{"location":"#","page":"Overview","title":"Overview","text":"In most wells, the fluid flow contains three distinct phases (oil, water, and gas) with temperature- and pressure-dependent properties, including density, viscosity. In addition, the fluids will exhibit varying degrees of miscibility and entrainment depending on pressure and temperature as well, such that the pressure change with respect to distance or depth along the wellbore is itself dependent on pressure:","category":"page"},{"location":"#","page":"Overview","title":"Overview","text":"fracph = f(pT)","category":"page"},{"location":"#","page":"Overview","title":"Overview","text":"Since the temperature profile varies according to depth, when assuming steady-state flow, consistent composition of each fluid, and a steady-state temperature profile uncoupled from pressure, the above can be re-expressed in terms of distance and pressure only:","category":"page"},{"location":"#","page":"Overview","title":"Overview","text":"fracdpdh = f(ph)","category":"page"},{"location":"#","page":"Overview","title":"Overview","text":"Note that f is composited from many empirical functions¹ and in most cases cannot be expressed in a tractable analytical form when dealing with multiphase flow. This is further complicated in gas lift wells, where the injection point is also pressure dependent, but changes the fluid composition and pressure profile above it.","category":"page"},{"location":"#","page":"Overview","title":"Overview","text":"Currently, no mechanistic or empirical correlations fully capture the variability in these three-phase fluid flows (or the direct properties of the fluids themselves), so the most performant methods are typically matched to the conditions and fluids they were developed to describe².","category":"page"},{"location":"#","page":"Overview","title":"Overview","text":"Most techniques for applying these correlations to calculate pressure profiles involve dividing the wellbore into a 1-dimensional series of discrete segments and calculating an average pressure change for the entire segment. Increasing segmentation will generally improve accuracy at the cost of additional computation. The most feasible and stable method for resolving the pressure change in each segment is typically some form of fixed-point iteration by
[…]
