### 2019-08-29 jnoynaert/PressureDrop.jl#3 @jnoynaert
@mancellin I've implemented the suggested changes:
- added Gadfly citation to paper
- added some very high level notes on the theory on [the documentation landing page](https://jnoynaert.github.io/PressureDrop.jl/stable/)
- added some file format examples in [the documentation](https://jnoynaert.github.io/PressureDrop.jl/stable/core/#Wellbores-1) as well as links to example files in the readme

Let me know if you have any other suggestions!
