### 2019-08-30 #1642 @mancellin
PressureDrop.jl is a specialized tool to model the pressure profile in oil and gas wells.
The code can be installed and tested easily using Julia package manager. It is well documented and the code behaved as documented when I tried it.

Some comments and suggestions have been discussed there: https://github.com/jnoynaert/PressureDrop.jl/issues/3.

Overall, I can recommend the paper for publication.
