### 2019-08-30 jnoynaert/PressureDrop.jl#3 @mancellin
Great!

The new overview in the documentation is very helpful. So, the big picture of your problem is actually a non-linear ODE. Maybe you should then check DifferentialEquations.jl instead of NLSolve.

One more thing: the use of non-SI units feels a bit weird to me. But I guess it is standard in your industry, isn't it?
