### 2019-08-30 jnoynaert/PressureDrop.jl#3 @jnoynaert
Non-SI units are indeed the standard in the U.S., even for academia. Outside the U.S. typical SI units are also common, but things can get stranger (reporting oil in metric tons in Russia, for example).

I'll look into using DiffEq.jl for the solution--that's an attractive idea because I think I might be able to formulate it so that I get a solution along the entire flow path in a single call instead of iterating by segment.

Thanks for the input!
