### 2019-09-18 #1642 @wkearn
Sorry, again, for taking so long with my review.

This package looks great! It is definitely designed with a specialized audience in mind, but the code is very clearly documented, and someone with experience using these kinds of models will, I think, have no problem running the models and analyzing the output.

I have a few minor suggestions that I will add to jnoynaert/PressureDrop.jl/issues/3 but otherwise I am happy to recommend this paper for publication.
