### 2019-09-18 jnoynaert/PressureDrop.jl#3 @wkearn
It seems like you added filepaths to the documentation, e.g.

```
valvefilepath = joinpath(dirname(dirname(pathof(PressureDrop))), "test/testdata/valvedata_wrappers_1.csv"
```

It would be nice if you duplicated that in the README. A potential user could then copy and paste the code example from the README and run it successfully.

The other suggestion I have would be a comparison between the model output from PressureDrop.jl and one of the commercial packages mentioned in the paper, like a plot showing that the output from PressureDrop.jl is more or less identical to the software people might otherwise use.
