### 2019-09-18 jnoynaert/PressureDrop.jl#3 @jnoynaert
Great ideas; I'll add both of those.
