### 2019-09-19 #1642 @labarba
@wkearn You said "otherwise" you're happy to recommend publication. Thus, we wait for you to confirm that your minor suggestion is addressed, and you are satisfied.
