### 2019-09-27 95b04a3
minor readme updates, added template for other software comparisons, fixed range syntax to preserve backwards compat with Julia 1.0
