+++ docs/src/similartools.md
@@ -1,0 +1,13 @@
+# Similar tools
+
+This package is intended as a free and flexible substitute for commercial tools.
+
+## Comparison
+
+
+| Syntax      | Description |
+| ----------- | ----------- |
+| Header      | Title       |
+| Paragraph   | Text        |
+
+Add linkback in readme.
