+++ docs/src/plotting.md
@@ -14,9 +14,9 @@ All plotting functions wrap Gadfly plot definitions.
 using PressureDrop
 
 segments = 100
-MDs = range(0, 5000, length = segments) |> collect
+MDs = range(0, stop = 5000, length = segments) |> collect
 incs = repeat([0], inner = segments)
-TVDs = range(0, 5000, length = segments) |> collect
+TVDs = range(0, stop = 5000, length = segments) |> collect
 
 well = Wellbore(MDs, incs, TVDs, 2.441)
 
