### 2019-10-07 #1642 @jnoynaert
Sorry, I haven't had a chance to get back and make those documentation changes yet. I should have some spare time this week to finish them.
