### 2019-10-18 jnoynaert/PressureDrop.jl#3 @jnoynaert
@wkearn I finally made the updates to the readme & documentation; see [here](https://jnoynaert.github.io/PressureDrop.jl/v1.0/similartools/) for a comparison table and a visual comparison to one of the more popular tools.

As far as numerical output comparisons go, a whole section of integration testing is validated against outputs from a commercial package (the other sections coming from reference texts)
