### 2019-10-18 #1642 @jnoynaert
@oliviaguest finally pushed this and tagged Will in the [original issue](https://github.com/jnoynaert/PressureDrop.jl/issues/3)
