### 2019-10-18 3929cb9
minor doc update
