### 2019-11-04 #1642 @wkearn
@oliviaguest Sorry for the delay on this, but @jnoynaert's updates look really good, and I can fully recommend publication.
