### 2019-11-04 #1642 @jnoynaert
@oliviaguest I just looked at the last article proof and it looks good to me. Thanks for your patience on this one!
