### 2019-11-05 9f00930
Update paper.bib
Fixing DOIs for paper
