+++ paper/paper.bib
@@ -45,7 +45,9 @@
   volume={102},
   number={1},
   year={1980},
-  publisher={American Society of Mechanical Engineers}
+  publisher={American Society of Mechanical Engineers},
+  doi={10.1115/1.3227845},
+  url={https://doi.org/10.1115/1.3227845}
 }
 
 @misc{Jones:2019,
@@ -82,6 +84,6 @@
   title        = {GiovineItalia/Gadfly.jl: v1.1.0},
   month        = jul,
   year         = 2019,
-  doi          = {10.5281/zenodo.3337394},
-  url          = {https://doi.org/10.5281/zenodo.3337394}
+  doi          = {10.5281/zenodo.593105},
+  url          = {https://doi.org/10.5281/zenodo.593105}
 }
