### 2019-11-05 #1642 @jnoynaert
@oliviaguest I noticed a few DOIs needed fixing, but the check references command for whedon is referencing an older version of the paper than what's on the master branch right now.

What do we need to do to sync up to the latest version for the reference checks? I'm wondering if it's stuck on the release archived on Zenodo
