### 2019-11-08 #1642 @kyleniemeyer
Hi @jnoynaert, just a few things before I finalize acceptance:

1. In the newest Zenodo record (10.5281/zenodo.3529746, I believe), can you fix the title to match that of the paper?

2. One of our [paper requirements](https://openjournals.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) is
> A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience.

Can you edit the introduction to make it a bit more accessible to a more general audience?

3. Also, a number of the older references are missing information:
- Payne:1979 reference: missing `doi={10.2118/6874-PA}`, `volume=31`, `number=9`
- Ramey:1962 reference: missing `doi={10.2118/96-PA}`, `volume=14`, `number=4`
- Griffith:1961 reference: missing `doi={10.1115/1.3682268}, `volume=83`, `number=3`, `pages={307--318}`
- Beggs:1973 reference: missing `doi={10.2118/4007-PA}`, `volume=25`, `number=5`
