### 2019-11-08 10f15fe
Added simplified paper intro for JOSS
