### 2019-11-08 286b691
Adding additional DOIs to paper
