+++ paper/paper.bib
@@ -11,7 +11,12 @@
   Title = "{Two-Phase Slug Flow}",
   journal = {J. of Heat Transfer},
   year = 1961,
-  month = aug
+  month = aug,
+  volume = 83,
+  number = 3,
+  pages = {307--318}
+  doi={10.1115/1.3682268},
+  url={https://doi.org/10.1115/1.3682268}
 }
 
 @article{Beggs:1973,
@@ -19,7 +24,11 @@
   Title = "{A Study of Two-Phase Flow in Inclined Pipes}",
   journal = {Journal of Petroleum Technology},
   year = 1973,
-  month = may
+  month = may,
+  volume = 25,
+  number = 5,
+  doi={10.2118/4007-PA},
+  url={https://doi.org/10.2118/4007-PA}
 }
 
 @article{Payne:1979,
@@ -27,7 +36,11 @@
   Title = "{Evaluation of Inclined-Pipe, Two-Phase Liquid Holdup and Pressure-Loss Correlations using Experimental Data}",
   journal = {Journal of Petroleum Technology},
   year = 1979,
-  month = sep
+  month = sep,
+  volume=31,
+  number=9,
+  doi={10.2118/6874-PA},
+  url={https://doi.org/10.2118/6874-PA}
 }
 
 @article{Ramey:1962,
@@ -35,7 +48,11 @@
   Title = "{Wellbore Heat Transmission}",
   journal = {Journal of Petroleum Technology},
   year = 1962,
-  month = apr
+  month = apr,
+  volume=14,
+  number=4,
+  doi={10.2118/96-PA},
+  url={https://doi.org/10.2118/96-PA}
 }
 
 @article{Shiu:1980,
