+++ paper/paper.bib
@@ -14,7 +14,7 @@
   month = aug,
   volume = 83,
   number = 3,
-  pages = {307--318}
+  pages = {307--318},
   doi={10.1115/1.3682268},
   url={https://doi.org/10.1115/1.3682268}
 }
