### 2019-11-08 #1642 @jnoynaert
@kyleniemeyer
- updated title @ [doi.org/10.5281/zenodo.3529746](url))
- added references and verified the last DOI-less citation does not have one available (at least a reasonably discoverable one, at any rate)
- added two introductory paragraphs to clarify a bit for people not working in that field
